Reorganize db enums, minor fixes
