+++ deepwell/src/models/relation.rs
@@ -1,8 +1,6 @@
 //! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.19
 
-use crate::services::relation::RelationType;
-
-use crate::services::relation::RelationObjectType;
+use crate::types::{RelationObjectType, RelationType};
 use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
