+++ deepwell/src/services/file_revision/structs.rs
@@ -20,8 +20,7 @@
 
 use super::prelude::*;
 use crate::services::page_revision::PageRevisionCountOutput;
-use crate::types::FetchDirection;
-use crate::types::FileRevisionType;
+use crate::types::{FetchDirection, FileRevisionType};
 
 #[derive(Debug, Clone)]
 pub struct CreateFileRevision {
