+++ deepwell/src/services/page_revision/structs.rs
@@ -19,8 +19,7 @@
  */
 
 use super::prelude::*;
-use crate::types::PageRevisionType;
-use crate::types::{FetchDirection, PageDetails, PageId};
+use crate::types::{FetchDirection, PageDetails, PageId, PageRevisionType};
 use ftml::layout::Layout;
 use ftml::parsing::ParseError;
 use std::num::NonZeroI32;
