+++ deepwell/src/services/user/structs.rs
@@ -21,8 +21,7 @@
 use super::prelude::*;
 use crate::models::alias::Model as AliasModel;
 use crate::models::user::Model as UserModel;
-use crate::types::Bytes;
-use crate::types::UserType;
+use crate::types::{Bytes, UserType};
 use std::net::IpAddr;
 use time::Date;
