+++ deepwell/src/types/enums/license.rs
@@ -1,5 +1,5 @@
 /*
- * types/enums.rs
+ * types/enums/license.rs
  *
  * DEEPWELL - Wikijump API provider and database manager
  * Copyright (C) 2019-2026 Wikijump Team
@@ -18,171 +18,9 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-//! Enums used across multiple services that correspond to database TEXT columns.
-
 use sea_orm::DeriveValueType;
 use serde::{Deserialize, Serialize};
 use strum_macros::{Display, EnumIter, EnumString};
-
-#[derive(
-    EnumIter,
-    Serialize,
-    Deserialize,
-    Debug,
-    Copy,
-    Clone,
-    Hash,
-    PartialEq,
-    Eq,
-    DeriveValueType,
-    EnumString,
-    Display,
-)]
-#[sea_orm(value_type = "String")]
-#[serde(rename_all = "kebab-case")]
-#[strum(serialize_all = "kebab_case", ascii_case_insensitive)]
-pub enum UserType {
-    Regular,
-    System,
-    Site,
-    Bot,
-}
-
-#[derive(
-    EnumIter,
-    Serialize,
-    Deserialize,
-    Debug,
-    Copy,
-    Clone,
-    Hash,
-    PartialEq,
-    Eq,
-    DeriveValueType,
-    EnumString,
-    Display,
-)]
-#[sea_orm(value_type = "String")]
-#[serde(rename_all = "kebab-case")]
-#[strum(serialize_all = "kebab_case", ascii_case_insensitive)]
-pub enum AliasType {
-    Site,
-    User,
-}
-
-#[derive(
-    EnumIter,
-    Serialize,
-    Deserialize,
-    Debug,
-    Copy,
-    Clone,
-    Hash,
-    PartialEq,
-    Eq,
-    DeriveValueType,
-    EnumString,
-    Display,
-)]
-#[sea_orm(value_type = "String")]
-#[serde(rename_all = "kebab-case")]
-#[strum(serialize_all = "kebab_case", ascii_case_insensitive)]
-pub enum FileRevisionType {
-    Regular,
-    Rollback,
-    Create,
-    Delete,
-    Undelete,
-    Move,
-}
-
-#[derive(
-    EnumIter,
-    Serialize,
-    Deserialize,
-    Debug,
-    Copy,
-    Clone,
-    Hash,
-    PartialEq,
-    Eq,
-    DeriveValueType,
-    EnumString,
-    Display,
-)]
-#[sea_orm(value_type = "String")]
-#[serde(rename_all = "kebab-case")]
-#[strum(serialize_all = "kebab_case", ascii_case_insensitive)]
-pub enum PageRevisionType {
-    Regular,
-    Rollback,
-    Undo,
-    Create,
-    Delete,
-    Undelete,
-    Move,
-}
-
-#[derive(
-    EnumIter,
-    Serialize,
-    Deserialize,
-    Debug,
-    Copy,
-    Clone,
-    Hash,
-    PartialEq,
-    Eq,
-    DeriveValueType,
-    EnumString,
-    Display,
-)]
-#[sea_orm(value_type = "String")]
-#[serde(rename_all = "kebab-case")]
-#[strum(serialize_all = "kebab_case", ascii_case_insensitive)]
-pub enum MessageRecipientType {
-    Regular,
-    Cc,
-    Bcc,
-}
-
-impl sea_orm::TryFromU64 for MessageRecipientType {
-    fn try_from_u64(_: u64) -> std::result::Result<Self, sea_orm::DbErr> {
-        Err(sea_orm::DbErr::ConvertFromU64(
-            "cannot construct MessageRecipientType from u64; auto_increment must be false for this primary key",
-        ))
-    }
-}
-
-#[derive(
-    EnumIter,
-    Serialize,
-    Deserialize,
-    Debug,
-    Copy,
-    Clone,
-    Hash,
-    PartialEq,
-    Eq,
-    DeriveValueType,
-    EnumString,
-    Display,
-)]
-#[sea_orm(value_type = "String")]
-#[serde(rename_all = "kebab-case")]
-#[strum(serialize_all = "kebab_case", ascii_case_insensitive)]
-pub enum TextBlockType {
-    Code,
-    Html,
-}
-
-impl sea_orm::TryFromU64 for TextBlockType {
-    fn try_from_u64(_: u64) -> std::result::Result<Self, sea_orm::DbErr> {
-        Err(sea_orm::DbErr::ConvertFromU64(
-            "cannot construct TextBlockType from u64; auto_increment must be false for this primary key",
-        ))
-    }
-}
 
 #[derive(
     EnumIter,
