+++ deepwell/src/types/mod.rs
@@ -22,7 +22,6 @@
 
 mod array;
 mod bytes;
-mod connection_type;
 mod conversion_error;
 mod enums;
 mod fetch_direction;
@@ -38,11 +37,11 @@ mod rerender_depth;
 
 pub use self::array::ArrayLength;
 pub use self::bytes::Bytes;
-pub use self::connection_type::ConnectionType;
 pub use self::conversion_error::{EnumConversionError, parse_layout};
 pub use self::enums::{
-    AliasType, FileRevisionType, License, MessageRecipientType, PageRevisionType,
-    TextBlockType, UserType,
+    Action, AliasType, ConnectionType, FileRevisionChange, FileRevisionType, License,
+    MessageRecipientType, PageRevisionChange, PageRevisionType, RelationObjectType,
+    RelationType, Resource, TextBlockType, UserType,
 };
 pub use self::fetch_direction::FetchDirection;
 pub use self::file_details::FileDetails;
@@ -51,6 +50,6 @@ pub use self::maybe::Maybe;
 pub use self::page_details::PageDetails;
 pub use self::page_id::PageId;
 pub use self::page_order::PageOrder;
-pub use self::permissions::{Action, Permission, Resource};
+pub use self::permissions::Permission;
 pub use self::reference::Reference;
 pub use self::rerender_depth::RerenderDepth;
